Revert "Update travis config file to rebase on master before test running"
This reverts commit bd99a90faa8188eb8c9ada6b68c403fff6a800c2.

While this is common in cockpit's projects, osbuild projects refrain
from rebasing branches.

Rebasing on master as it happens to be when the test is running, leaves
you in a state where you neither know that the branch itself is working,
nor that it will work when you actually do merge it later, because
master might have changed in between.