devel: add full-stack development environment
This integrates all the Image Builder components needed by the
cloud.redhat.com frontend and allows them to be developed and run
locally using `docker compose`.

This should make it simple to make patches across the different
components and develop them in tandem.

Thanks to Achilleas Koutsou for the initial idea and implementation
in osbuild-composer.

Signed-off-by: Tom Gundersen <teg@jklm.no>