CreateImageWizard: update azure redirect
On authorization of image builder on azure, the user will be redirected
to portal.azure.com. The response mode is changed to query instead of
form_post because form_post makes a POST request while query makes a GET
request. portal.azure.com does not respond when we make a POST request
to it so users will not be redirected unless we use the response mode
of query.