Fix docker stack (#713)
* devel: update composer config

Update some of the config for the docker-compose
file since there have been a few changes and the
config is slightly out of date. Additionally, a
refresh_token parameter was added to the oauth_url
since this is used by composer's mock openid provider
to generate the `rh-id-org` key in the jwt token.

* distribution: update frontend entrypoint

Update the command for spinning up the frontend
to match the new script command.

* devel: remove offline token

Remove the offline token and replace
with a client secret instead.

* devel: add metrics profile

Only launch the metrics containers when
explictly requested. This helps decrease
the logging output when the metrics are
not required. The metrics can be called
with `docker-compose --profile metrics up`