Wizard: Update activation keys link text and URI
This commit updates the link on the registration step so users are
directed to the exact place where they can manage keys.

Because managing keys through the customer portal will soon become
deprecated (keys will be managed through the keys service on insights)
the text was also changed to be more general.